various tweaks to the optimizer for the simulation.
Fixed some bugs that were causing incorrect handling of the first & last days of trading within the simulation. Changed the metric used for evaluating whether a number of historical days is actually a good one.

Looks like I should be using as much history as possible. Having a smaller amount of history does not significantly impact the accuracy of the final price predictions, but does significantly impact the quality of the resulting distribution